split stuff up into stuff
