+++ entities.py
@@ -1,0 +1,9 @@
+class Player:
+    def __init__(self):
+        self.playerpos = 0
+        self.playerhp = 10
+
+class Monster:
+    def __init__(self):
+        self.hp = 10
+        self.atk = 10
