+++ rooms.py
@@ -1,0 +1,3 @@
+class Room:
+    def __init__(self):
+        pass
